1、所有弹幕类分开文件，避免文件内容臃肿，易于拓展 2、获取弹幕接口支持type参数，指定获取的数据类型，支持：xml、json，默认获取json类型
