+++ UpdateLog.md
@@ -1,4 +1,12 @@
-﻿# 2025-04-29
+﻿# 2025-05-14
+1. 所有接口使用蓝图统一管理，避免app.py文件内容臃肿
+2. 所有弹幕类分开文件，避免文件内容臃肿，易于拓展
+3. 获取弹幕接口支持type参数，指定获取的数据类型，支持：xml、json，默认获取json类型
+
+# 2025-05-09
+1. 修复爱奇艺get_episode_url只能获取第一个键值的BUG
+
+# 2025-04-29
 一. 修改获取弹幕时平台链接的获取逻辑：
 1. 先查询数据库内数据，如果没有数据进行2
 2. 如果是非第一季或者douban_id是空值，通过豆瓣小程序进行匹配
