+++ app.py
@@ -4,7 +4,7 @@ from flask import Flask
 
 from apps.api import api_app
 from apps.danmu.app import danmu
-from apps.fnos.app import fnos
+from apps.fnos.app import fnos_app
 from core.db.db import engine
 from core.db.model import Base
 
@@ -12,7 +12,7 @@ app = Flask(__name__)
 app.logger.setLevel(logging.INFO)
 
 app.register_blueprint(danmu, url_prefix='/danmu')
-app.register_blueprint(fnos, url_prefix='/fnos')
+app.register_blueprint(fnos_app, url_prefix='/fnos')
 app.register_blueprint(api_app, url_prefix='/api')
 
 # 创建表
